Log Drains - Poll new (for real), restart, and deprovision/delete
Polling was not previously working right after a new log drain was
created because it was "pending" and not "provisioning".

A new component provides a Restart and Deprovision action on
provisioned log drains, only the Deprovision action on provisioning /
pending log drains, and no actions on a deprovisioning log drain.

Seeing a deprovisioning log drain in the UI is not likely to happen
since sweetness deletes it once the operation is complete and we will
remove it from the local store. Subsequent calls to fetch all log
drains will no longer include it once the operation is complete.