Add support for clustered databases
On the backend, clustered databases and replicated databases behave the
same way (i.e. initialize-from), but on the frontend it's worth creating
some distinctions in how we represent them and encourage users to use
them.

None of the limitations imposed here are necessary for things to work on
the backend, but they should result in a better user experience on the
frontend.