One select to rule them all,
One select to find them, one select to bring them all and in the
darkness bind them.

This commit consolidates all the various kinds of selects that we
have previously had into one single component - ultimate-select.

(The name can be debated and of course changed, but in lieu of a
better one..)

The ultimate-select basically works the same way as the former
primitive-select and object-select did. To work with object
collections, you choose the key and value by specifiying itemKey
and itemValue respectively. This works with both POJOs and Ember
objects. For primitive types, don't specify itemKey and itemValue
and it should work just like primitive-select.

AFAICS this also removes all usages of Ember views, and now we're
just using components \o/