Create and delete default vhosts
This commit adds the ability to create and delete default vhosts from
the dashboard.

This depends on aptible/api.aptible.com#307 and aptible/sweetness#525
and hence this cannot be merged until those PRs are merged.

As for the UX itself, the radio button is only revealed if the service
does not already have a default vhost set on it. An outstanding
question is how to only display this option for stacks that have
default vhosts enabled (there are a bunch of production stacks which
have this enabled as well as all(?) shared stacks). Or should this be
enabled for all stacks?

Is the radio button clear enough for customers to know what it's
about? Is there any way it could be made clearer?